consolidate run on tick
