+++ boards/DCM/Src/App/states/App_FaultState.c
@@ -1,3 +1,4 @@
+#include "states/App_AllStates.h"
 #include "states/App_FaultState.h"
 
 #include "App_SharedMacros.h"
@@ -12,18 +13,8 @@ static void FaultStateRunOnEntry(struct StateMachine *const state_machine)
 
 static void FaultStateRunOnTick(struct StateMachine *const state_machine)
 {
-    struct DcmWorld *world = App_SharedStateMachine_GetWorld(state_machine);
-    struct DcmCanTxInterface *can_tx      = App_DcmWorld_GetCanTx(world);
-    struct DcmCanRxInterface *can_rx      = App_DcmWorld_GetCanRx(world);
-    struct BrakeLight *       brake_light = App_DcmWorld_GetBrakeLight(world);
 
-    bool is_brake_actuated =
-        App_CanRx_FSM_BRAKE_GetSignal_BRAKE_IS_ACTUATED(can_rx);
-    bool is_regen_active =
-        App_CanTx_GetPeriodicSignal_TORQUE_REQUEST(can_tx) < 0.0f;
-
-    App_BrakeLight_SetLightStatus(
-        brake_light, is_brake_actuated, is_regen_active);
+    App_AllStatesRunOnTick(state_machine);
 }
 
 static void FaultStateRunOnExit(struct StateMachine *const state_machine)
