+++ boards/DCM/Src/App/states/App_InitState.c
@@ -1,3 +1,4 @@
+#include "states/App_AllStates.h"
 #include "states/App_InitState.h"
 #include "states/App_DriveState.h"
 
@@ -13,18 +14,7 @@ static void InitStateRunOnEntry(struct StateMachine *const state_machine)
 
 static void InitStateRunOnTick(struct StateMachine *const state_machine)
 {
-    struct DcmWorld *world = App_SharedStateMachine_GetWorld(state_machine);
-    struct DcmCanTxInterface *can_tx      = App_DcmWorld_GetCanTx(world);
-    struct DcmCanRxInterface *can_rx      = App_DcmWorld_GetCanRx(world);
-    struct BrakeLight *       brake_light = App_DcmWorld_GetBrakeLight(world);
-
-    bool is_brake_actuated =
-        App_CanRx_FSM_BRAKE_GetSignal_BRAKE_IS_ACTUATED(can_rx);
-    bool is_regen_active =
-        App_CanTx_GetPeriodicSignal_TORQUE_REQUEST(can_tx) < 0.0f;
-
-    App_BrakeLight_SetLightStatus(
-        brake_light, is_brake_actuated, is_regen_active);
+    App_AllStatesRunOnTick(state_machine);
 
     // No need for any safety checks, just run! (this is a demo)
     App_SharedStateMachine_SetNextState(state_machine, App_GetDriveState());
